[type-system] deduplicate fieldTypeToLlvm and add fail-loud default
