+++ src/codegen/statements/control-flow.ts
@@ -38,13 +38,7 @@ import {
 } from "../infrastructure/ir-builders.js";
 import { SymbolKind_Number, SymbolKind_String } from "../infrastructure/symbol-table.js";
 import type { FieldInfo } from "../infrastructure/type-resolver/types.js";
-import {
-  stripOptional,
-  classifyArray,
-  ArrayKind_None,
-  arrayKindToLlvm,
-  fieldTypeToLlvmPrimitive,
-} from "../infrastructure/type-system.js";
+import { stripOptional, fieldTypeToLlvm } from "../infrastructure/type-system.js";
 import { setWantsI1 } from "../expressions/condition-generator.js";
 import { tryOptimizeWhileLoopMap } from "./loop-idiom.js";
 
@@ -834,7 +828,7 @@ export class ControlFlowGenerator {
     for (let i = 0; i < commonFields.length; i++) {
       const f = commonFields[i] as CommonField;
       keys.push(stripOptional(f.name));
-      types.push(this.fieldTypeToLlvm(f.type));
+      types.push(this.fieldTypeToLlvmLocal(f.type));
       tsTypes.push(f.type);
     }
 
@@ -864,25 +858,9 @@ export class ControlFlowGenerator {
     return type;
   }
 
-  private fieldTypeToLlvm(fieldType: string): string {
-    const prim = fieldTypeToLlvmPrimitive(fieldType);
-    if (prim) return prim;
-    const ak = classifyArray(fieldType);
-    if (ak !== ArrayKind_None) return arrayKindToLlvm(ak);
-    if (fieldType.startsWith("Map<"))
-      return fieldType.startsWith("Map<string,") ? "%StringMap*" : "%Map*";
-    if (fieldType === "Set<string>") return "%StringSet*";
-    if (fieldType.startsWith("Set<")) return "%Set*";
+  private fieldTypeToLlvmLocal(fieldType: string): string {
     if (this.isEnumType(fieldType)) return "double";
-    if (fieldType.indexOf(" | ") !== -1) {
-      const parts = fieldType.split(" | ");
-      for (let i = 0; i < parts.length; i++) {
-        const part = parts[i].trim();
-        if (part === "null" || part === "undefined") continue;
-        return this.fieldTypeToLlvm(part);
-      }
-    }
-    return "i8*";
+    return fieldTypeToLlvm(fieldType);
   }
 
   private isEnumType(typeName: string): boolean {
@@ -982,7 +960,7 @@ export class ControlFlowGenerator {
     for (let i = 0; i < ifaceAllFields.length; i++) {
       const f = ifaceAllFields[i] as { name: string; type: string };
       keys.push(stripOptional(f.name));
-      types.push(this.fieldTypeToLlvm(f.type));
+      types.push(this.fieldTypeToLlvmLocal(f.type));
       tsTypes.push(f.type);
     }
